Start moving deployment functionality to core
