+++ resources/plugins/QuantME-CamundaPlugin/deployment/services/DeploymentPlugin.js
@@ -18,7 +18,7 @@ import ServiceDeploymentInputModal from './ServiceDeploymentInputModal';
 import ServiceDeploymentBindingModal from './ServiceDeploymentBindingModal';
 import { getRootProcess } from 'client/src/app/quantme/utilities/Utilities';
 
-import { createServiceInstance, uploadCSARToContainer } from './OpenTOSCAUtils';
+import { createServiceInstance, uploadCSARToContainer } from 'client/src/app/quantme/deployment/OpenTOSCAUtils';
 import { bindUsingPull, bindUsingPush, getBindingType } from './BindingUtils';
 import { getServiceTasksToDeploy } from './DeploymentUtils';
